+++ packages/cldr-schema/src/schema/general/types.ts
@@ -4,7 +4,7 @@ import { ContextTransformFieldType } from './autogen.context';
 
 export type LineOrderType = 'ltr' | 'rtl';
 export type CharacterOrderType = 'ttb' | 'btt';
-export type ContextType = 'begin-sentence' | 'standalone' | 'ui-list-or-menu';
+export type ContextType = 'middle-of-text' | 'begin-sentence' | 'standalone' | 'ui-list-or-menu';
 
 export interface LayoutSchema {
   readonly characterOrder: FieldArrow;